+++ src/fwriteR.c
@@ -65,6 +65,7 @@ writer_fun_t *funs[] = {
   &writeDateInt32,
   &writeDateFloat64,
   &writePOSIXct,
+  &writePOSIXctInt,
   &writeNanotime,
   &writeString,
   &writeCategString,
@@ -124,11 +125,11 @@ static int32_t whichWriter(SEXP column) {
   case LGLSXP:
     return logical01 ? WF_Bool32 : WF_Bool32AsString;
   case INTSXP:
-    if (INHERITS(column, char_POSIXct))    return WF_POSIXct;
     if (isFactor(column))                  return WF_CategString;
     if (dateTimeAs == DATETIMEAS_EPOCH)    return WF_Int32;
     if (INHERITS(column, char_ITime))      return WF_ITime;
     if (INHERITS(column, char_Date))       return WF_DateInt32;
+    if (INHERITS(column, char_POSIXct))    return WF_POSIXctInt;
     return WF_Int32;
   case REALSXP:
     if (INHERITS(column, char_nanotime)